requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
jinja2
bcrypt
pydantic[email]